moved spp to detection modules
